+++ core/bootstrap/shutdown.js
@@ -1,9 +1,17 @@
+const logger = require("../utils/logger");
+
 module.exports = function registerShutdown(server) {
   const shutdown = (signal) => () => {
     logger.warn(`Received ${signal}, shutting down server...`);
-    server.close(() => {
-      process.exit(0);
-    });
+
+    if (server && typeof server.close === "function") {
+      server.close(() => {
+        process.exit(0);
+      });
+      return;
+    }
+
+    process.exit(0);
   };
 
   process.on("SIGINT", shutdown("SIGINT"));